d'abord trouver le type, puis essayer l'overflow. je vais essayer de check l'overflow dans les fonctions de check

+++ repo/ex00/inc/ScalarConverter.hpp
@@ -11,7 +11,7 @@ class ScalarConverter
 {
 	private :
 		static char		_char;
-		static long		_int;
+		static int		_int;
 		static float	_float;
 		static double	_double;
